Fix default param value for keyframe effects with factor. Fixes #18 Fix wrong version of effect used because of unfaithful double comparison

+++ src/assets/abstractassetsrepository.hpp
@@ -69,7 +69,7 @@ protected:
         QString id;    // identifier of the asset
         QString mltId; //"tag" of the asset, that is the name of the mlt service
         QString name, description, author, version_str;
-        double version;
+        int version;
         QDomElement xml;
         AssetType type;
     };
